VideoProperties: return fps as a float type
Double type is not constructed the same way in C++ and Java, and does
not exist in python. Our bindings of 'C++ double' could create issues,
and since a fps with a simple precision is enough, we decided to return
fps as float.

+++ src/AvTranscoder/mediaProperty/VideoProperties.hpp
@@ -56,7 +56,7 @@ public:
 	int getProfile() const;
 	int getLevel() const;
 
-	double getFps() const;
+	float getFps() const;
 
 	bool hasBFrames() const;
 	//bool isClosedGop() const;
